qemu: Fix cleanups in case vm or remote was not initialized

+++ qemu/boot.py
@@ -24,5 +24,9 @@ class BootTest(test.VirtTest):
         self.vm.login_remote()
 
     def cleanup(self):
-        self.vm.remote.run('shutdown -h now')
-        self.vm.power_off()
+        if self.vm:
+            if self.vm.remote:
+                self.vm.remote.run('shutdown -h now')
+                # TODO: Wait for machine to go down
+            self.vm.power_off()
+
